refactor: move LabelNotFounndError to separated file
