+++ src/errors/LabelNotFoundError.ts
@@ -1,0 +1,1 @@
+export class LabelNotFoundError extends Error {}
